requirements.txt
# Core runtime
numpy>=1.24.0,<2.0.0
PyYAML>=6.0
tqdm>=4.65.0
langid>=1.1.6

# Tokenization and chunking
chonkie>=1.0.2,<1.1.0
tiktoken>=0.5.0
tokenizers>=0.15.0

# Hugging Face ecosystem
huggingface-hub>=0.36.0
transformers>=4.40.0
sentence-transformers>=2.2.0
safetensors>=0.4.3
Pillow>=9.4.0

# PyTorch stack (install GPU-enabled build manually if needed)
torch>=2.0.0

# Retrieval & indexing
faiss-gpu>=1.8.0
bm25s[core]==0.2.1

# Datasets & preprocessing
datasets[http]>=2.18.0
pandas>=2.0.0
pyarrow>=12.0.0

# Evaluation & utilities
scikit-learn>=1.3.0
